refactor: factor our getGuests() and move to lib/api/guest
