+++ src/lib/api/guest.ts
@@ -12,3 +12,11 @@ export async function getGuest(guestId: number): Promise<GuestDataAPIResponse> {
     (await response.body.json()) as GuestDataAPIResponse;
   return guestResponse;
 }
+export async function getGuests() {
+  const response = await API.post({
+    apiName: "auth",
+    path: "/getGuests",
+  }).response;
+  const guestsResponse = (await response.body.json()) as GuestsAPIResponse;
+  return guestsResponse;
+}
